CMake build system integration (#161)
* Initial CMake build

* CommonGateway build with cmake

* Added cellular in cmake POC

* Add testing to CMake POC + latest gtest and fff.h libs

* Disabled not working tests per platform

* Add specification of CXX standard for linux builds

* Remove obsolete files and fix a test compilation issue

* Add Doxygen generation with CMake

* Switch to GTest library in master branch

* Separate core libs in their own cmake files and remove NT user from fff library comments

* Add quickstart guide with CMake
Remove unused compiler flags and generalize freertos paths

* Address comments from Sebastian's review

Added comments in CMake files
Removed unnecessary statements from some unittests and added feature checks
Reverted changes to Queue and Logger unit tests - they are separate tasks
Updated Jenkinsfile
Updated quick-start guide with examples for building using CMake
Cleaned-up bstlib CMakeLists
Renamed toolchain.cmake -> toolchain_arm_gcc.cmake
Moved some common variables from board_configs to root CMake file

* Update Jenkinsfile to call cmake properly

* Update Jenkinsfile to call cmake properly

* Explicitly specify current directory as source for CMake in docker

* Remove functions not available in CMake 3.7 (the one in docker toolchain)

* Remove unavailable command argument NAME_WLE

* Attempt to fix build of unittests

* Take care for git case-sensitivity

* Attempt to archive test output

* Fix undefined behavior in CmdProcessor unit tests

* Enable more tests and enable warnings and debug symbols for tests

* Added proper feature checks to all unittests, which actually enabled some of them. Excluded SPI, I2C and UART tests from build as they need attention to work properly.

* Remove old Doxygen file

+++ core/utils/test/unit/source/TLV_unittest.cc
@@ -131,7 +131,7 @@ protected:
             /* calculate group Handle Size based on the base address of the application Buffer
              * and the address of the created first Element
              */
-            groupHandleSize = (uint32_t) Element - (uint32_t) DataBuffer;
+            groupHandleSize = (uint64_t) Element - (uint64_t) DataBuffer;
         }
         else
         {
